% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_indices.R
\name{calculate_indices}
\alias{calculate_indices}
\title{Calculate humification and biological indices (HIX and BIX)}
\usage{
calculate_indices(x)
}
\arguments{
\item{x}{A FEEM tibble generated by \code{read_feem}.}
}
\value{
A tibble with columns \code{param} and \code{value}.
}
\description{
Calculate humification and biological indices (HIX and BIX)
}
\examples{
file <- list.files(
path = system.file("extdata", package = "cwrshelpr"),
full.names = TRUE,
pattern = ".+\\\\.csv"
)
data <- read_feem(file[1])
calculate_indices(data)
}
