% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate_regions.R
\name{integrate_regions}
\alias{integrate_regions}
\title{Fluorescent regional integration (FRI)}
\usage{
integrate_regions(
  x,
  ex_min = c(200, 200, 200, 250, 250),
  ex_max = c(250, 250, 250, 340, 400),
  em_min = c(200, 330, 380, 200, 380),
  em_max = c(330, 380, 550, 380, 550),
  region_names = 1:5
)
}
\arguments{
\item{x}{An excitation-emission matrix generated by cwrshelpr::read_feem().}

\item{ex_min}{Lower excitation bound of each region.}

\item{ex_max}{Upper excitation bound of each region.}

\item{em_min}{Lower emission bound of each region.}

\item{em_max}{Upper emission bound of each region.}

\item{region_names}{A vector of region names, corresponding to ex/em_min/max.}
}
\value{
A tibble with columns \code{name}, \code{ex_min}, \code{ex_max}, \code{em_min}, \code{em_max}, \code{integrated}.
}
\description{
Fluorescent regional integration (FRI)
}
\examples{
file <- list.files(
   path = system.file("extdata", package = "cwrshelpr"),
   full.names = TRUE,
   pattern = ".+\\\\.csv"
)
feem <- read_feem(file[1], truncate = TRUE)
integrate_regions(feem)
}
