% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_lims.R
\name{read_lims}
\alias{read_lims}
\title{Read AGAT LIMS data files}
\usage{
read_lims(
  path,
  sheet_name = 1,
  first_line = "AGAT Workorder",
  meta_rows = 4,
  work_order
)
}
\arguments{
\item{path}{Path to stored ICP-MS files}

\item{sheet_name}{Name of MS Excel sheet where data are stored}

\item{first_line}{The first column matches this pattern at the first line of data.
As an input to \code{stringr::str_detect()}, this can be a partial match or a regex.}

\item{meta_rows}{How many rows contain metadata that should appear in the column names? Default is 4.}

\item{work_order}{The AGAT LIMS work order number. Used to select columns containing numeric data.}
}
\value{
A tibble with columns \code{param}, \code{unit}, \code{rdl} (detection limit), \code{value},
\code{date}, \code{bdl} (below detection indicator), and \code{sample}
}
\description{
Read AGAT LIMS data files
}
\examples{
file <- list.files(
   path = system.file("extdata", package = "cwrshelpr"),
   full.names = TRUE,
   pattern = ".+\\\\.xlsx"
)
read_lims(file[3], work_order = "00A123456")
}
