% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_icp.R
\name{read_icp}
\alias{read_icp}
\title{Read ICP-MS data files generated by the Thermo Scientific iCAP-RQ}
\usage{
read_icp(
  path,
  sheet_name = "Dilution factor included",
  first_line = "Sample List"
)
}
\arguments{
\item{path}{Path to stored ICP-MS files}

\item{sheet_name}{Name of MS Excel sheet where data are stored}

\item{first_line}{The first column matches this pattern at the first line of data.
As an input to \code{stringr::str_detect()}, this can be a partial match or a regex.}
}
\value{
A tibble with columns \code{sample_name}, \code{category}, \code{isotope}, \code{element},
\code{value}, and \code{unit}
}
\description{
Read ICP-MS data files generated by the Thermo Scientific iCAP-RQ
}
\examples{
file <- list.files(
   path = system.file("extdata", package = "cwrshelpr"),
   full.names = TRUE,
   pattern = ".+\\\\.xlsx"
)
read_icp(file[1])
}
